refactor: move Android Glance adapters into components
Keep Voltra Glance component adapters in a flat components package and preserve fixed lazy grid column counts above five.
