+++ packages/android-client/android/src/main/java/voltra/glance/components/ButtonRenderers.kt
@@ -1,4 +1,4 @@
-package voltra.glance.renderers
+package voltra.glance.components
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.graphics.Color
@@ -15,6 +15,9 @@ import androidx.glance.layout.Box
 import androidx.glance.unit.ColorProvider
 import voltra.glance.LocalVoltraRenderContext
 import voltra.glance.applyClickableIfNeeded
+import voltra.glance.renderers.RenderNode
+import voltra.glance.renderers.extractImageProvider
+import voltra.glance.renderers.getOnClickAction
 import voltra.glance.resolveAndApplyStyle
 import voltra.models.VoltraElement
 import voltra.styling.JSColorParser
