+++ packages/android-client/android/src/main/java/voltra/glance/components/ProgressRenderers.kt
@@ -1,4 +1,4 @@
-package voltra.glance.renderers
+package voltra.glance.components
 
 import androidx.compose.runtime.Composable
 import androidx.glance.GlanceModifier
