+++ packages/android-client/android/src/main/java/voltra/glance/components/VoltraLazyVerticalGrid.kt
@@ -1,20 +1,19 @@
-package voltra.glance.renderers
+package voltra.glance.components
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.unit.dp
 import androidx.glance.GlanceModifier
 import androidx.glance.appwidget.lazy.GridCells
-import androidx.glance.appwidget.lazy.LazyColumn
 import androidx.glance.appwidget.lazy.LazyVerticalGrid
-import androidx.glance.layout.Alignment
 import voltra.glance.LocalVoltraRenderContext
 import voltra.glance.applyClickableIfNeeded
+import voltra.glance.renderers.RenderNode
 import voltra.glance.resolveAndApplyStyle
 import voltra.models.VoltraElement
 import voltra.models.VoltraNode
 
 @Composable
-fun RenderLazyColumn(
+fun VoltraLazyVerticalGrid(
     element: VoltraElement,
     modifier: GlanceModifier? = null,
 ) {
@@ -29,67 +28,11 @@ fun RenderLazyColumn(
             element.t,
             element.hashCode(),
         )
-
-    // Extract props
-    val horizontalAlignment = extractHorizontalAlignment(element.p)
-
-    LazyColumn(
-        modifier = finalModifier,
-        horizontalAlignment = horizontalAlignment,
-    ) {
-        when (val children = element.c) {
-            is VoltraNode.Array -> {
-                items(children.elements.size) { index ->
-                    RenderNode(children.elements[index])
-                }
-            }
-
-            is VoltraNode.Ref -> {
-                val resolved = context.sharedElements?.getOrNull(children.ref)
-                if (resolved is VoltraNode.Array) {
-                    items(resolved.elements.size) { index ->
-                        RenderNode(resolved.elements[index])
-                    }
-                } else {
-                    item { RenderNode(resolved) }
-                }
-            }
-
-            null -> { /* Empty list */ }
-
-            else -> {
-                item { RenderNode(children) }
-            }
-        }
-    }
-}
-
-@Composable
-fun RenderLazyVerticalGrid(
-    element: VoltraElement,
-    modifier: GlanceModifier? = null,
-) {
-    val context = LocalVoltraRenderContext.current
-    val (baseModifier, _) = resolveAndApplyStyle(element.p, context.sharedStyles)
-    val finalModifier =
-        applyClickableIfNeeded(
-            modifier ?: baseModifier,
-            element.p,
-            element.i,
-            context.widgetId,
-            element.t,
-            element.hashCode(),
-        )
-
-    // Extract grid configuration from props
-    val gridCells = extractGridCells(element.p)
-
-    val horizontalAlignment = extractHorizontalAlignment(element.p)
 
     LazyVerticalGrid(
-        gridCells = gridCells,
+        gridCells = extractGridCells(element.p),
         modifier = finalModifier,
-        horizontalAlignment = horizontalAlignment,
+        horizontalAlignment = extractHorizontalAlignment(element.p),
     ) {
         when (val children = element.c) {
             is VoltraNode.Array -> {
@@ -118,18 +61,10 @@ fun RenderLazyVerticalGrid(
     }
 }
 
-private fun extractHorizontalAlignment(props: Map<String, Any?>?): Alignment.Horizontal =
-    when (props?.get("horizontalAlignment") as? String) {
-        "start" -> Alignment.Horizontal.Start
-        "center-horizontally" -> Alignment.Horizontal.CenterHorizontally
-        "end" -> Alignment.Horizontal.End
-        else -> Alignment.Horizontal.Start
-    }
-
 private fun extractGridCells(props: Map<String, Any?>?): GridCells =
     when (val columns = props?.get("columns")) {
         is Number -> {
-            GridCells.Fixed(columns.toInt().coerceIn(1, 5))
+            GridCells.Fixed(columns.toInt().coerceAtLeast(1))
         }
 
         is String -> {
